+++ app/src/main/java/net/lachlanmckee/bookmark/service/model/Folder.kt
@@ -1,7 +1,7 @@
 package net.lachlanmckee.bookmark.service.model
 
 data class Folder(
-  val id: Int,
-  val parentId: Int?,
+  val id: Long,
+  val parentId: Long?,
   val name: String
 )
